requirements.txt
bottle
gitpython
mercurial